Refactor into folders for Bookcase standards, comments.
