+++ Bookcase/Patches/StardewValley/Tools/FishingRod/FishCaughtPatch.cs
